requirements.txt
PyYAML==6.0
requests==2.26.0
duckdb==0.3.4
pytest==7.1.2
fastapi==0.78.0
uvicorn==0.17.6
fastparquet==0.8.1
pyarrow==8.0.0
tqdm==4.64.0
numpy==1.21.6
pandas==1.3.5
pytest-asyncio==0.19.0
anyio==3.6.2